Move location of raw access logs
This is a data project which means it's going to work with a lot of
data. I originally included a reference access log in the main directory
with the expectation that any scripts should be able to run correctly
against that reference log (and without the intent of including real
access logs in the source control going forward).

Now that we've started working with real data though, we've changed the
way that we reference it in our scripts (access_log.*) so that we can
include all dates in the folder and handle gzipped copies. It would be
nice to organize all of the raw input data into a separate folder and
then when we transform/anonymize that data store the outputs in yet
another folder. In some cases, we may want to publish/share those
transformed/anonymized outputs so that people can see the end reports
that we generate with this project.

To accomplish this, I'm going to add the reference access log into the
raw folder, make sure all of the scripts still work, and then ignore
that folder going forward after moving the real logs into it locally.