Build an interactive scatterplot to visualize the PCA results
I wanted an easier way to visualize the relative locations of the food
images as they were mapped in the output of the PCA onto a
multidimensional space. Earlier we were just comparing the left most 10
foods with the right most foods on each dimension. That worked well
enough, but it didn't really tell us how close they were to each other
on those dimensions. I wanted to see the various mapped dimensions two
at a time on a scatter plot with easy access to see where images fell on
that scatter plot (either showing the images themselves on the plot, or
giving an easy hover over access to see the images). Ideally, I also
wanted to be able to chose which dimension I was looking at on each
axis, and for fun, add in a slower transition when changing the
dimension used so that it would look a bit like a rotation (allowing the
user to see specific points move).