Add support for switch over to GCP
We're moving the hunger service over from AWS to GCP which means the
logs will be handled slightly differently. We'll still be using the old
logs for a while, this step just added a new data prep version that can
read in the new ones. We still need some logic to put both sets
together.