Create/cleanup resources for food and tag tree explorers
I've been playing around with some d3.js pages for exploring the
clusters generated by our R script that generates the crystal bowl
decision trees. We have a lot of tags that overlap in meaning though and
the same tools can be used with some slight modifications to get a
better feel for our food tags. So far everything has just run on my
machine though -- in order to share the explorers with the whole team
I'm just cleaning everything up before pushing a copy to this
repositories docs folder so that it can be hosted on github.