Increase weighting for tags that are known foods/ingredients
I've looked for a few sources of public datasets that include food and
or ingredient lists. A lot of our tags seem to be irrelevant and
completely unrelated to food. Others fall somewhere in between --
cultures, countries or nationalities for instance aren't themselves
food, but are definitely relevant when it comes to categorizing styles
of cooking.

Short of manually going through the tag list and cleaning it up, I
wanted to see if there was any easy way to automate the process. One
approach would be to use the word embeddings themselves to try to train
a model of food related/unrelated terms (that may be an interesting next
project).

The approach that I took for now though was to find some sources of
ingredients which would clearly include terms that are food related,
then cross reference those terms with our list of tags and finally,
rather than limit out tags to the food related ones, at the very least
increase the weighting given to known food related terms.