Add a list of tags manually flagged as irrelevant
This list will act as our initial stop tag list to see how it effects
the food space. A number of the tags are definitely food related, but
not necessarily helpful to our goal of mapping out the food images --
food in particular is a clear "stop word" as it goes without saying
considering the context of our scope. Some terms like meal, plate, bowl,
table may be indirectly relevant. This is just a starting point, the
list will definitely be tuned.