Handle multiword tags
For better or worse, I've added in logic to explode multi-word tags into
all possible ngram cobinations -- so "cuba libre" becomes three tags:
cuba, libre and cubalibre. This meant that we now potentially have some
new tags that we didn't include in the original creation of our filtered
dictionary, so I went ahead and created a new filtered dictionary along
with a new improved process for generating it (it runs a lot faster than
the old one even if it takes a bit more hand holding). See notes in the
filter-dictionary.R script included in this commit.

The food clusters getting created at this point are looking very
unbalanced, but they at least seem to be placing similar foods more or
less in the same viscinity. In general this seems to suggest that the
approach is likely putting foods into a meaningful meaningspace. I think
at this point we just need to switch over to a different approach to
building our decision trees.