Make decision tree more forgiving
When users exclude an entire branch of food based on a single click of
an image, then the decision tree becomes less tolerant of bad image
pairings. The left branch may still contain a lot of food that the user
wants even though they happened to prefer the image that represented the
right branch. This is more of an issue in the beginning of the session
as each click quickly excludes large swaths of food.

There's definitely a better way to implement such fault tolerance on the
tree navigation side, but the quickest and easiest way to implement it
for now was to build redundant trees. I tried out a couple variations
and found in the end that we want to avoid too much overlap --
particularly as we get closer to the terminal nodes -- as it starts to
lead to frequently pairing the same food up against itself -- as well as
huge trees that don't add much value. I've also doubled up on the first
two dimensions of most significant differentiation as their multipliers
in the SVD diagonal were nearly twice that of some of the ones that
follow close behind.