Evaluate imagenet CNN based approach to building food clusters
So far we've been building our foodspace based on food image tags. One
down side to this is that somebody has to tag the food images first.
Another downside is that the tagging approach has to be somewhat
consistent. A lot of the images that we've scraped from image sharing
sites aren't tagged with our food app in mind and sometimes they aren't
even in the same language.

Given that our whole concept is very visually oriented -- the idea that
you will know what you when when you see a picture of it -- it could be
interesting to see if we can use a Convolutional Neural Network to build
our food space -- and more specifically if it would be as easy as taking
an off the shelf model pretrained on imagenet and using it to score our
food images without any retraining.

This was my initial attempt -- just a little hacking around -- it's not
exactly what we want, but it shows some promise. Ultimately I think
we'll need to do some transfer learning -- we can probably take a
pretrained model from imagenet, drop one or two fully connected layers
on top of it, train those layers specifically with some food image data
and then use that new model to map images into a meaningful food space.