Update image metadata references
We refactored the location of some of the files in the images repository
when we moved over to using firebase. The local d3 js files need to
reference the new locations for them to work both locally as well as the
ones that are published on github pages (those have already been updated
though).