Cleanup
This project started out as a single directory with a number of scripts.
It got a bit messy and when things changed, rather than cleaning up the
main directory, I created a new directory for a specific purpose and
started writing new scripts there. Then I did the same for another
purpose, and again and so on. At this point, all of the scripts from the
main directory are sadly out of date and aside from using them as
cut/paste resources they haven't been touched in ages. To avoid any
major confusion, I'm just deleting all of those scripts right now --
they'll always still be available in git's history if needed, but it was
time for a good cleanup. This isn't perfect, but at least now rather
than having one big mess, we have lots of little messes...