Evaluate fasttext as alternative to GloVe
The glove word embeddings definitely worked better than nothing, but we
struggled with the problem that a lot of our tags weren't in the
pretrained dictionary of 400K words. Fasttext offers a pretrained
dictionary with more words, but also a model that can generate
embeddings for words that aren't in it's dictionary by building the
embedding from parts of the word and words that they match. I don't know
exactly how it works, but fasttext comes with a tool that allows you to
generate custom word or sentence vectors for a list of input
words/sentences from it's own command line tool. It requires an extra
step -- generate lists of sentences (titles) and words (tags), then run
those lists through the fasttext command line and suddenly we have our
own custom dictionary of word/sentence vectors. This is very similar to
the speedup step we did with the glove embeddings where we parsed down
the full dictionary of 400k embeddings to only the thousand or so that
we actually used. In the end, it's pretty easy to create a drop in
replacement for what we were already using with better coverage of our
tagspace.

Qualitatively (based on tag similarities showing up in the word-tree
explorer), the embeddings don't look particularly worse or better than
the GloVe ones did. They do seem to handle typos fairly well (mistyped
words tend to show up as nearest neighbors to their correctly typed
word). Compound tags could use some work -- for now I've been creating
single word tags from compound word tags by changing spaces to hyphens.
The compound words often show up as each others nearest neighbors
though, even if they aren't particularly similar for any apparent
reason. I wonder if it's the hyphen itself that's causing the similarity
in these cases. We may want to try alternate approaches simply removing
the spaces rather than swapping them over to hyphens.