Implement tag based clustering
Upgrade the new fasttext based approach to work with tags instead of
just titles. I still need to figure out the best looking approach for
combining the three different types of tags, but to start out it's
helpful just to evaluate how well each one is working individually. In
the process I saw that the vectors we were getting back from fasttext
weren't always unit vectors, so I've added in a function to normalize
the vectors in our matrices. I'm also renormalizing after averaging the
tag vectors for a tag type. Next we'll need to add in a weigting scheme
for adding the different groups together.

On the clustering side, I've added in chunks of logic for taking the
food space matrix and building dendogram trees using hierarchical
agglomerative clustering as well as random projectins and our previous
approach of PCA.