Build and save consistent projection matrix
We've been using svd to build lower dimension projections of our higher
dimension vector space. One of the problems with the approach is that we
have a fairly stable model for our word vectors -- aside from switching
over from glove to fasttext, we're still use a 300 dimension vector
space that doesn't change from run to run (we don't retrain it). Our
food vectors move around a bit as we fix the tags or the way that we
weight the tags, but the space that the foods around in are still stable
-- each of the 300 dimensions doesn't change in its meaning. The
moment that we use SVD to project our foods down onto a lower
dimensional space though, we chose a projection that maximized the
variance in our food sample -- thus it depends on the foods in the
sample, and their location in the space. It would be nice if we could
rely on a stable projection mechanism. I considered using random
projections (with a random seed), but they were quite difficult to have
any confidence in. I've also been interested in building our own hand
picked dimensions using anchor foods and some QR decomposition, but I
recently realized that we might be able to accomplish something similar
if we simply run SVD to build the projection matrix using a fixed set of
anchor foods. We don't have to use all of the foods, but rather a
smaller sample that's chosen for a good amount of variance (not too much
redundancy with some of the same foods repeated over and over, and
avoiding outliers with foods whose tags haven't been edited yet).

This commit introduces that new projection matrix (saving the matrix
itself just in case we have any trouble reproducing it) along with some
newer d3 javascript visualization. Firefox stopped supporting CORS
requests for file type URLs, so I've kind of been forced to ditch file
urls and run a static server everytime I want to view the files locally.
Fortunately that's pretty easy if you have npm installed -- just run
`npx http-server` from the d3 folder then navigate to localhost:8080.