Update word scatter plot with new projection
Now that we have an approach for building a reusable projection matrix
for projecting our 300 dimension meaning space down onto a lower
dimensional space, I wanted to update the word analysis script to use
that same projection so that we can see tags plotted on the same set of
dimensions as we're using for plotting foods. Eventually I'd like to
bring the two visuals together -- it would be nice if you could click on
a food in the scatter plot and have the tags for that food show up as
points with a different color.