Update everything to use the same indexed tree
I've been supporting a couple different food tree formats up to this
point and it's gotten to be a bit of a headache. I'd like to publish the
newest data explorers with foods from adams apple, but my local copy is
out of sync with what's published. Getting everything over to the same
format will make things a lot easier going forward.