Remove r scripts for old glove based approach
We've decided to make a hard break and stop supporting the old approach
to building feed me jefferson decision trees. We can always go back to
an old commit if necessary, but in general we've decided that the way
forward to the next release of feed me jefferson is through building out
the adams apple platform. That means a newer/more consistent food space
representation with indexed btrees rather than the old mess of sometimes
a pure node/children structure, but one of many arbitrary variations
taylored specifically for a data explorer. Focusing on the newer
approach means we can clean up and consolidate a lot of code that's
gotten out of hand.