Serialize metadata that supports food tree
I started to go down a deep rabit hole trying to build these as
configurably sized partial food basket files, but decided it might be
best to leave that implementation to a javascript library since the
client that has to read in the partials and put them back together will
also ultimately be written in javascript. That code may chose to take
the attribution data from a different source, but I've included that as
an output in addition to the vector files which can only be produced
here because R is where they originate.