Fix transitions in data explorer
After adding the colors to denote whether a food image had been edited
or whether it was still going off of the original tags, I ran into the
same problem that I hit a while ago when I tried to add in a feature to
make it easier to find specific food ids in the tree -- if works
initially, but when you click on one of the nodes the effects don't
transition correctly to the new node locations. Last time around I
couldn't figure out what the problem was -- this time I saw it was
because I was applying the colors to the node element itself rather than
the circle child element of the node. Now that I figured that out I was
also able to go back and add in the search feature -- so now if you
include a query param like ?id=0000129 at the end of you data explorer
page url it will highlight the food you're looking for by coloring it
red and making it larger.