Bug fix
Caught a bug in two of the methods that were referring to
cluster/clusters interchangeably -- one happened to work because in
older cases because it was a globally available variable, but when name
of the variable in the main script was changed, it no longer worked. It
should have been referring to the argument rather than the globally
available variable name. Updated and fixed.