Create script to reindex exported moderator meta data
I've already started using this in practice. It still needs to be
cleaned up a bit, but I want to commit what I have been using for
historical reasons.

It builds both the moderator indexes and the json files to support the
data explorers (for both foods and tags).  We still need to evaluate the
best way of building the vectors themselves, but this script makes it
more clear how to aggregate tag vectors for a food and potentially build
food vectors on the fly (which could be nice for the moderator).

This doesn't includes the logic to build the decision trees -- that
probably needs to be separated out into a different script.