Migrate github pages from docs folder to gh-pages branch
I'd rather we not increase the size of the repo every time we publish
changes to the data explorers just because some metadata has changed.
It's a lot easier to clean up a secondary branch that we never merge in
than to prune history out of master.

R scripts will still expect that the docs folder is there like it used
to be, but we'll ignore it from a git perspective and treat it as a
separately managed git repo folder. I could set it up as a submodule,
but those get kind of messy and this is just for managing the pages.