Refactor folder structure
I'm moving github pages to a branch and creating a new assets repository
for managing our moderator asset export/imports. I've updated the newer
scripts to point to a different set of symlinked folders as a result so
that they read from and write directly out to the appropriate locations.