Ignore resources used for local hosting
I'm using `npx http-server -p 9000` to quickly standup a local http
server for testing the site pages locally. There are plenty of other
options, but this fits well with a lot of the other javascript we do
elsewhere on feedme jefferson. With npx I don't really have to have the
package installed locally, but I don't like to download it everytime I
want to run the server.