Clean up html/styles -- start of refactoring
There's a lot of repetition here. It would be nice to get rid of it.
Maybe eventually the html pages which are for the most part identical
will get rebuilt as some kind of a buildable web app, but for now I'm
fine with the way it is. Mostly I want to get rid of the redundancy
between the different javascript files.