Refactor and add new features
I bundled a few things together in this commit. Mainly I was trying to
refactor the code to get rid of redundancy between the various
d3 javascript implementations. We have two types of visuals -- flare
dendograms vs scatter plots and two types of content -- tags vs food
images. Aside from those differences, the bulk of the logic is the same
across the files -- as are many of the convenience features like
highlighting a food by adding an id to the URL query params to make it
easier to find. I've added in some generalized support for query string
params to make it easier to add new features in the future. I also added
support for bookmarking directly into an initial set of dimensions on
scatter plots, and finally made an attempt at some better mobile support
for single clicking as a replacement for hovering (along with locking to
avoid unhovering when you're using a mouse and want to go over to the
metadata for the item you've hovered over without losing it because
you had to hover over other items to get to it -- now you can click the
item to lock it on scatter plots).