More features -- highlights, sampling, alt data, color
I've inconsistently added a few features.

Data: In order to compare different data sets, I've added support
for a user provided "data" query param that specifies where to
pull the input data file from.

Highlights: all visualizers now support a comma separated list of
elements to highlight rather than limiting the highlight to a single
value.

Filters/Sampling: The scatter plots can be kind of dense, particularly
when viewing tags as text using "show=words". To limit the number of
tags/foods that show up on the scatter plot, you can filter out ranges
by dimension with a query string parameter like X4=40,60 (that would
exclude all elements in the range between the scaled 40/60 percent on
that dimension) and/or you can provide a random sampling percent with
the query param sample=10 (which would show approximately 10 percent of
the already filtered elements).

Color support for visualizing the third dimension on scatter plots. I've
only started to implement this for the food plot. I need to add in a
legend and fixtures for selecting which dimension to apply it to, but
for now you can select it on the query param list with z=X4 (or any
other dimension name).