Add support for alternate dictionaries
I'm playing around with designer dimensions based on anchor terms. Some
of the terms don't exist in our tag set or they might go missing. I'd
rather create a separate dictionary based on a hand set of curated terms
that may or may not exist as tags and have an easy way of pulling in
that fasttext tag vector as a separate file.