Try out a new method for generating projection matrix
This isn't entirely reproducible as it's based on the set of tags pulled
in from the moderator -- this will change over time, so it's important
to document here that it was based off of the assets repository commit

382bee4611e10bd92760c48b42b474a2f54ecd9c

in case we ever need to reproduce it. You'd need to run the script
against the tags that were available at that time. Running it against
new tags might produce better results, but we might need to modify some
of the numbers in the script to get similar outcomes.