Play around with some concepts
I wanted to look into building custom "designer dimensions" rather than
relying on running SVD against a curated set of data. I think there's
some promise there, but it needs some more work, and we'd have to come
up with a few good ideas for dimensions.

I also had one of those middle of the night thoughts that keeps you up
at night -- food cubes (or dice). It basically provides us with an
alternate way of packaging up baskets of food -- fewer food items at a
time -- while maximizing those foods ability to target a set of
dimensions in callibrating the users appetite. It will take more work on
the back end, and quite a bit of rework on the front end as well, but in
the end, it would leave more of the logic on the back end and give us a
simpler set of logic on the frontend. The back end logic would be mostly
offline providing us with the ability to build a mostly static site with
limited "randomness" thrown in from the front end to provide the user
with enough variety to keep it from gettin boring.